Added functions & refactor

+++ src/string/ft_strlen.c
@@ -1,10 +1,10 @@
 #include "libft.h"
 
-int ft_strlen(const char *str)
+size_t ft_strlen(const char *str)
 {
 	if (str == NULL)
 		return 0;
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i] != '\0')
